+++ nix/tests.nix
@@ -43,7 +43,8 @@ let
         type != "directory"
         || hasPrefix "${dir}/" "${relPath}/"
         || hasPrefix "core-stub/" "${relPath}/"
-        || hasPrefix "base-stub/" "${relPath}/";
+        || hasPrefix "base-stub/" "${relPath}/"
+        || hasPrefix "json-stub/" "${relPath}/";
       src = ../test;
       name = "test-${dir}-src";
     };